+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## ChangeLog
 
+### 1.0.5
+* Added stopRefresh method
+* Added refreshPhoto method
+* Added restore method
+* Updated the readme
+* Code cleaned
+
 ### 1.0.0
 * Initial release